fix(pgtap): correct plan count + sum(bigint) cast in cost-dashboard tests
CI pgtap caught two test-only bugs the DATA agent couldn't run locally (no Supabase):
- usage_summary_cache_cols.test.sql: sum(bigint) returns numeric in Postgres, so
  is(numeric, 1000::bigint) errored (function is(numeric,bigint) does not exist) and
  aborted 2 of 8 subtests → cast the sum to ::bigint.
- agent_run_stats.test.sql: plan(12) but there are exactly 11 assertions (8 AC-ACD-005
  + 3 AC-ACD-006, all passing) → plan(11).

RPCs + expectations unchanged (the 11/6 assertions that ran all passed); these are
plan/type-cast fixes only.

+++ supabase/tests/usage_summary_cache_cols.test.sql
@@ -57,7 +57,7 @@ select is(
   50::bigint, 'AC-ACD-003 chat Σreasoning_tokens = 50 (owner-scoped)');
 -- own-org isolation: the org Y 9999s never enter A1's totals.
 select is(
-  (select coalesce(sum(cached_tokens),0) from org_usage_summary()),
+  (select coalesce(sum(cached_tokens),0)::bigint from org_usage_summary()),
   1000::bigint, 'AC-ACD-003 org-Admin cached total excludes other orgs (own-org only)');
 reset role;
